fix shiny odds not growing, changes to many maps, much more

+++ include/pokedex.h
@@ -34,7 +34,7 @@ struct PokedexEntry
 
 void ResetPokedex(void);
 u16 GetPokedexHeightWeight(u16 dexNum, u8 data);
-u16 GetNationalPokedexCount(u8);
+u32 GetNationalPokedexCount(u8);
 u16 GetHoennPokedexCount(u8);
 u8 DisplayCaughtMonDexPage(u16 dexNum, u32 otId, u32 personality);
 s8 GetSetPokedexFlag(u16 nationalNum, u8 caseId);
